First attempt at setting the lint settings with lerna
