+++ packages/provisioning/package.json
@@ -1,30 +1,22 @@
 {
-  "name": "atat-serverless-poc",
+  "name": "atat-portfolio-draft-api",
   "version": "1.0.0",
-  "description": "ATAT Serverless framework PoC",
+  "description": "ATAT Portfolio Draft API",
   "scripts": {
     "test": "jest --verbose --no-watchman",
     "start": "func host start",
     "lint": "eslint --ext .js,.ts,.json .",
-    "lint:fix": "eslint --ext .js,.ts,.json --fix ."
+    "lint:fix": "npm run lint -- --fix"
   },
   "keywords": [
     "azure",
     "serverless"
   ],
   "devDependencies": {
-    "@azure/functions": "^1.2.3",
     "@types/jest": "^26.0.23",
     "@types/node": "^15.12.5",
-    "@typescript-eslint/eslint-plugin": "^4.28.2",
-    "@typescript-eslint/parser": "^4.28.2",
     "cache-loader": "^4.1.0",
     "eslint": "^7.30.0",
-    "eslint-config-standard": "^16.0.3",
-    "eslint-plugin-import": "^2.23.4",
-    "eslint-plugin-json": "^3.0.0",
-    "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^5.1.0",
     "fork-ts-checker-webpack-plugin": "^6.2.12",
     "jest": "^27.0.6",
     "serverless-azure-functions": "^2.1.0",
@@ -35,5 +27,11 @@
     "typescript": "^4.3.5",
     "webpack": "^5.41.1",
     "webpack-node-externals": "^3.0.0"
+  },
+  "dependencies": {
+    "@azure/cosmos": "^3.12.1",
+    "@azure/functions": "^1.2.3",
+    "dayjs": "^1.10.6",
+    "uuid": "^8.3.2"
   }
 }
