+++ packages/provisioning/shared_infrastructure/terraform.tfvars.template
