+++ packages/provisioning/src/lib/models/PortfolioSummary.ts
@@ -1,0 +1,6 @@
+import { ProvisioningStatus } from './ProvisioningStatus'
+import { BaseDocument } from './BaseDocument'
+
+export interface PortfolioSummary extends BaseDocument {
+    status: ProvisioningStatus
+}
