+++ packages/random_quote/.eslintrc.json
@@ -1,0 +1,35 @@
+{
+  "env": {
+    "node": true,
+    "es6": true,
+    "jest": true,
+    "es2020": true
+  },
+  "extends": [
+    "standard",
+    "plugin:@typescript-eslint/recommended",
+    "plugin:json/recommended"
+  ],
+  "parser": "@typescript-eslint/parser",
+  "parserOptions": {
+    "ecmaVersion": 12,
+    "sourceType": "module"
+  },
+  "plugins": [
+    "@typescript-eslint"
+  ],
+  "ignorePatterns": ["**/*.js", "coverage/**"],
+  "rules": {
+    "no-console": "warn"
+  },
+  "overrides": [
+    {
+      "files": [
+        "**/__tests__/*.{j,t}s?(x)"
+      ],
+      "env": {
+        "jest": true
+      }
+    }
+  ]
+}
