+++ packages/random_quote/lib/logging.js
