+++ packages/random_quote/serverless.common.yml
