+++ packages/random_quote/shared_infrastructure/modules/keyvault/outputs.tf
