+++ packages/random_quote/shared_infrastructure/modules/keyvault/variables.tf
