+++ packages/random_quote/shared_infrastructure/terraform.tfvars.template
