+++ packages/random_quote/shared_infrastructure/variables-auth.tf
