+++ packages/random_quote/shared_infrastructure/variables-keyvault.tf
