+++ packages/random_quote/shared_infrastructure/variables-network.tf
