+++ packages/random_quote/src/handlers/get.ts
