+++ packages/random_quote/src/util/quotes_helper.ts
