+++ packages/random_quote/webpack.config.js
