--- provisioning/src/lib/models/PortfolioFull.ts
@@ -1,6 +1,0 @@
-import {PortfolioSummary} from "./PortfolioSummary";
-import {PortfolioStep} from "./PortfolioStep";
-
-export interface PortfolioFull extends PortfolioSummary {
-    portfolio: PortfolioStep
-}
