--- provisioning/src/lib/models/PortfolioSummary.ts
@@ -1,6 +1,0 @@
-import {ProvisioningStatus} from "./ProvisioningStatus";
-import {BaseDocument} from "./BaseDocument";
-
-export  interface PortfolioSummary extends BaseDocument {
-    status: ProvisioningStatus
-}
