--- provisioning/src/lib/models/ProvisioningStatus.ts
@@ -1,6 +1,0 @@
-export const enum ProvisioningStatus {
-    NotStarted = "not_started",
-    InProgress = "in_progress",
-    Failed = "failed",
-    Completed = "completed"
-}
